Removed code for supporting build tags and debug variables
