+++ Dockerfile
@@ -19,8 +19,7 @@ COPY version/ version/
 
 # Build
 ARG LD_FLAGS
-ARG BUILD_TAGS
-RUN CGO_ENABLED=0 GOOS=linux go build -ldflags="$LD_FLAGS" ${BUILD_TAGS:+-tags "$BUILD_TAGS"} -a -o manager cmd/main.go
+RUN CGO_ENABLED=0 GOOS=linux go build -ldflags="$LD_FLAGS" -a -o manager cmd/main.go
 
 # Use distroless as minimal base image to package the manager binary
 # Refer to https://github.com/GoogleContainerTools/distroless for more details
