+++ controllers/argocd/policyrule.go
@@ -5,8 +5,6 @@ import (
 
 	"golang.org/x/mod/semver"
 
-	"github.com/argoproj-labs/argocd-operator/controllers/argoutil"
-
 	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
 	"github.com/argoproj-labs/argocd-operator/common"
 
@@ -132,7 +130,7 @@ func policyRuleForRedisHa(client client.Client) []v1.PolicyRule {
 }
 
 func policyRuleForDexServer() []v1.PolicyRule {
-	dexPolicyRules := []v1.PolicyRule{
+	return []v1.PolicyRule{
 		{
 			APIGroups: []string{
 				"",
@@ -148,16 +146,6 @@ func policyRuleForDexServer() []v1.PolicyRule {
 			},
 		},
 	}
-	// if dex storage override is enabled, add policy rules for storing dex related custom resources
-	// in kubernetes storage.
-	if argoutil.IsDexKubernetesStorageEnabled() {
-		dexPolicyRules = append(dexPolicyRules, v1.PolicyRule{
-			APIGroups: []string{"dex.coreos.com"},
-			Resources: []string{"*"},
-			Verbs:     []string{"*"},
-		})
-	}
-	return dexPolicyRules
 }
 
 func policyRuleForServer(cr *argoproj.ArgoCD) []v1.PolicyRule {
